2023.04.14 commit - header > selectbox를 component화 - footer 수정 등
